+++ docs/.vuepress/config.js
@@ -10,6 +10,7 @@ const versionWithLowDashs = version.replace(/\./g, "_")
 const slugMap = {
     "Common Prefix Ambiguities": "COMMON_PREFIX",
     "None Unique Grammar Name Found": "UNIQUE_GRAMMAR_NAME",
+    "Terminal Token Name Not Found": "TERMINAL_NAME_NOT_FOUND",
     "No LINE_BREAKS Error": "LINE_BREAKS",
     "Unexpected RegExp Anchor Error": "ANCHORS",
     "Token can never be matched": "UNREACHABLE",